docs: include example of React.lazy for tree-shaking
