+++ examples/react/large-form/src/App.tsx
@@ -1,9 +1,10 @@
+import { Suspense } from 'react'
 import { PeoplePage } from './features/people/page.tsx'
 
 export default function App() {
   return (
-    <div>
+    <Suspense fallback={<p>Loading...</p>}>
       <PeoplePage />
-    </div>
+    </Suspense>
   )
 }
