+++ examples/react/large-form/src/components/text-fields.tsx
@@ -1,0 +1,25 @@
+import { useStore } from '@tanstack/react-form'
+import { useFieldContext } from '../hooks/form-context.tsx'
+
+export default function TextField({ label }: { label: string }) {
+  const field = useFieldContext<string>()
+
+  const errors = useStore(field.store, (state) => state.meta.errors)
+
+  return (
+    <div>
+      <label>
+        <div>{label}</div>
+        <input
+          value={field.state.value}
+          onChange={(e) => field.handleChange(e.target.value)}
+        />
+      </label>
+      {errors.map((error: string) => (
+        <div key={error} style={{ color: 'red' }}>
+          {error}
+        </div>
+      ))}
+    </div>
+  )
+}
